backend/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
cssselect==1.2.0
cssselect2==0.7.0
Django==5.0.7
django-ckeditor==6.7.1
django-ckeditor-5==0.2.13
django-filter==24.3
django-js-asset==2.2.0
django-jsignature==0.12
django-signature==0.3.1
djangorestframework==3.16.0
html5lib==1.1
idna==3.10
lxml==5.2.2
mysqlclient==2.2.4
oscrypto==1.3.0
pillow==10.4.0
pycparser==2.22
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
pypdf==5.0.1
pyquery==2.0.0
python-bidi==0.6.3
python-http-client==3.3.7
PyYAML==6.0.2
qrcode==8.0
reportlab==4.2.5
requests==2.32.3
sendgrid==6.11.0
six==1.16.0
sqlparse==0.5.0
starkbank-ecdsa==2.2.0
svglib==1.5.1
tinycss2==1.3.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.3
webencodings==0.5.1
xhtml2pdf==0.2.16
